isolation_levels/requirements.txt
prettytable==2.4.0
psycopg2-binary==2.9.2
PyMySQL==1.0.2
python-dotenv==0.19.2
wcwidth==0.2.5
